优化特种弹药箱相关

+++ Content/Projectiles/AmmoBoxs/BaseAmmoBox.cs
@@ -13,13 +13,14 @@ namespace CalamityOverhaul.Content.Projectiles.AmmoBoxs
         public override string Texture => CWRConstant.Item + "Placeable/NapalmBombBox";
 
         private bool onProj;
+        private bool onDorp;
         protected int maxFrameNum = 1;
         protected Vector2 drawOffsetPos;
         public int FromeThisTImeID;
 
         public override void SetDefaults() {
             Projectile.width = 46;
-            Projectile.height = 46;
+            Projectile.height = 36;
             Projectile.tileCollide = false;
             Projectile.ignoreWater = true;
             SetAmmoBox();
@@ -66,12 +67,19 @@ namespace CalamityOverhaul.Content.Projectiles.AmmoBoxs
         public virtual void ClockFrame() => CWRUtils.ClockFrame(ref Projectile.frame, 10, maxFrameNum - 1);
 
         public virtual void Dorp() {
-            for (int i = 0; i < 9; i++) {
+            for (int i = 0; i < 13; i++) {
                 Tile tile = CWRUtils.GetTile(Projectile.Bottom / 16);
                 if (!(tile.HasTile && (Main.tileSolid[tile.TileType] || Main.tileSolidTop[tile.TileType]))) {
                     Projectile.position.Y += 1;
                 }
                 else {
+                    if (!onDorp) {
+                        for (int z = 0; z < 33; z++) {
+                            int stompDust = Dust.NewDust(Projectile.BottomLeft, Projectile.width, 4, DustID.Smoke, 0f, 0f, 100, default, 1.5f);
+                            Main.dust[stompDust].velocity *= 0.2f;
+                        }
+                        onDorp = true;
+                    }
                     break;
                 }
             }
@@ -98,14 +106,19 @@ namespace CalamityOverhaul.Content.Projectiles.AmmoBoxs
                 if (player.CWR().TryGetInds_BaseFeederGun(out BaseFeederGun gun)) {
                     gun.ShootCoolingValue = gun.CanRightClick ? 10 : 2;//因为和一些枪械的右键功能按键冲突，所以要额外设置一个长一些的时间
                 }
-                if (player.PressKey(false)) {
-                    Item item = player.ActiveItem();
+                if (player.PressKey(false) && Projectile.IsOwnedByLocalPlayer()) {
+                    Item item = player.GetItem();
                     if (CanClick(item)) {
                         Preprocessing(player, item);
                         if (ClickBehavior(player, item.CWR())) {
                             Projectile.Kill();
                         }
                     }
+                    else {
+                        player.QuickSpawnItem(Projectile.FromObjectGetParent(), new Item(FromeThisTImeID));
+                        Projectile.Kill();
+                    }
+                    return;
                 }
             }
             Dorp();
@@ -113,8 +126,9 @@ namespace CalamityOverhaul.Content.Projectiles.AmmoBoxs
 
         public override bool PreDraw(ref Color lightColor) {
             Texture2D value = CWRUtils.GetT2DValue(Texture);
-            Main.EntitySpriteDraw(value, Projectile.Center - Main.screenPosition + drawOffsetPos, CWRUtils.GetRec(value)
-                , onProj ? Color.Gold : lightColor, Projectile.rotation, CWRUtils.GetOrig(value) - new Vector2(0, 2), Projectile.scale, SpriteEffects.None, 0);
+            Vector2 drawPos = Projectile.BottomLeft - new Vector2(0, value.Height) - Main.screenPosition;
+            Main.EntitySpriteDraw(value, drawPos, CWRUtils.GetRec(value)
+                , onProj ? Color.Gold : lightColor, Projectile.rotation, Vector2.Zero, Projectile.scale, SpriteEffects.None, 0);
             return false;
         }
     }
